Added the common-lang dependency
